+++ javascript/active_element.js
@@ -1,0 +1,10 @@
+let activeElement
+
+export default {
+  get element () {
+    return activeElement
+  },
+  set (element) {
+    activeElement = element
+  }
+}
